requirements.txt
asgiref==3.2.10
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.8
django-codemirror2==0.2
django-cors-headers==3.4.0
django-cte==1.1.4
django-debug-toolbar==2.2
django-extensions==3.0.2
django-import-export==2.2.0
django-js-asset==1.2.2
django-mptt==0.9.1
django-reversion==3.0.7
django-river==3.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
drf-yasg==1.17.1
et-xmlfile==1.0.1
idna==2.10
inflection==0.5.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
mock==4.0.2
odfpy==1.4.1
openpyxl==3.0.4
packaging==20.4
pycodestyle==2.6.0
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
river-admin==0.6.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
